The tetrominoes are always places from left to right in the beginning. Can you adapt the algorithm to mix this up. Don't use random for this (keep it deterministic). Also don't randomize the starting positions (should remain center).
That's not what I meant. The problem is that when laying the first row all tetrominoes are placed (now) from right to left. I want there to be some variance. For instance, first tetromino is placed all the way to the right, then one to the left, then one center, then right again. Etc.