Add a menu with a cog icon (svg) in the top left corner. This dropdown menu (animated) allows user to switch to New Year's Countdown or Clock mode. It also allows for increasing the speed. It should be subtle.
Changing the settings in the menu should also update the params in the url.

Changing to New Year's should restart the animation.

It should also be possible to toggle sound in the menu.

Move the cog to the right side, remove the icon for sound, that's now toggled in menu.